# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Flare by TK Restaurant API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  sortOrder: number;
  isActive: boolean;
  itemCount?: number;
  createdAt: string;
}

export interface CategoryInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  slug: string;
  description?: string;
  imageUrl?: string;
  sortOrder?: number;
  isActive?: boolean;
}

export interface CategoryUpdate {
  /** @minLength 1 */
  name?: string;
  slug?: string;
  description?: string;
  imageUrl?: string;
  sortOrder?: number;
  isActive?: boolean;
}

export interface MenuItem {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
  categoryId: number;
  /** @nullable */
  categoryName?: string | null;
  price: number;
  /** @nullable */
  originalPrice?: number | null;
  /** @nullable */
  imageUrl?: string | null;
  isAvailable: boolean;
  isFeatured: boolean;
  /** @nullable */
  calories?: number | null;
  tags?: string[];
  createdAt: string;
}

export interface MenuItemInput {
  /** @minLength 1 */
  name: string;
  slug: string;
  description?: string;
  categoryId: number;
  price: number;
  originalPrice?: number;
  imageUrl?: string;
  isAvailable?: boolean;
  isFeatured?: boolean;
  calories?: number;
  tags?: string[];
}

export interface MenuItemUpdate {
  /** @minLength 1 */
  name?: string;
  slug?: string;
  description?: string;
  categoryId?: number;
  price?: number;
  originalPrice?: number;
  imageUrl?: string;
  isAvailable?: boolean;
  isFeatured?: boolean;
  calories?: number;
  tags?: string[];
}

export type DealDiscountType = typeof DealDiscountType[keyof typeof DealDiscountType];


export const DealDiscountType = {
  percentage: 'percentage',
  fixed: 'fixed',
  combo: 'combo',
} as const;

export interface Deal {
  id: number;
  title: string;
  slug: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  discountType: DealDiscountType;
  /** @nullable */
  discountValue?: number | null;
  /** @nullable */
  originalPrice?: number | null;
  /** @nullable */
  dealPrice?: number | null;
  isActive: boolean;
  sortOrder?: number;
  createdAt: string;
}

export type DealInputDiscountType = typeof DealInputDiscountType[keyof typeof DealInputDiscountType];


export const DealInputDiscountType = {
  percentage: 'percentage',
  fixed: 'fixed',
  combo: 'combo',
} as const;

export interface DealInput {
  /** @minLength 1 */
  title: string;
  slug: string;
  description?: string;
  imageUrl?: string;
  discountType: DealInputDiscountType;
  discountValue?: number;
  originalPrice?: number;
  dealPrice?: number;
  isActive?: boolean;
  sortOrder?: number;
}

export type DealUpdateDiscountType = typeof DealUpdateDiscountType[keyof typeof DealUpdateDiscountType];


export const DealUpdateDiscountType = {
  percentage: 'percentage',
  fixed: 'fixed',
  combo: 'combo',
} as const;

export interface DealUpdate {
  /** @minLength 1 */
  title?: string;
  slug?: string;
  description?: string;
  imageUrl?: string;
  discountType?: DealUpdateDiscountType;
  discountValue?: number;
  originalPrice?: number;
  dealPrice?: number;
  isActive?: boolean;
  sortOrder?: number;
}

export interface OrderItem {
  menuItemId: number;
  name: string;
  price: number;
  quantity: number;
}

export type OrderOrderType = typeof OrderOrderType[keyof typeof OrderOrderType];


export const OrderOrderType = {
  delivery: 'delivery',
  pickup: 'pickup',
} as const;

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  new: 'new',
  preparing: 'preparing',
  ready: 'ready',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export interface Order {
  id: number;
  customerName: string;
  customerPhone: string;
  /** @nullable */
  customerAddress?: string | null;
  orderType: OrderOrderType;
  status: OrderStatus;
  totalAmount: number;
  /** @nullable */
  specialInstructions?: string | null;
  items: OrderItem[];
  createdAt: string;
}

export type OrderInputOrderType = typeof OrderInputOrderType[keyof typeof OrderInputOrderType];


export const OrderInputOrderType = {
  delivery: 'delivery',
  pickup: 'pickup',
} as const;

export interface OrderItemInput {
  menuItemId: number;
  quantity: number;
}

export interface OrderInput {
  /** @minLength 1 */
  customerName: string;
  /** @minLength 1 */
  customerPhone: string;
  customerAddress?: string;
  orderType: OrderInputOrderType;
  specialInstructions?: string;
  items: OrderItemInput[];
}

export type OrderStatusUpdateStatus = typeof OrderStatusUpdateStatus[keyof typeof OrderStatusUpdateStatus];


export const OrderStatusUpdateStatus = {
  new: 'new',
  preparing: 'preparing',
  ready: 'ready',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export interface OrderStatusUpdate {
  status: OrderStatusUpdateStatus;
}

export type DashboardStatsOrdersByStatus = {
  new?: number;
  preparing?: number;
  ready?: number;
  completed?: number;
  cancelled?: number;
};

export interface DashboardStats {
  totalOrders: number;
  todayOrders: number;
  totalRevenue: number;
  todayRevenue: number;
  ordersByStatus: DashboardStatsOrdersByStatus;
  recentOrders: Order[];
}

export type ListMenuItemsParams = {
categoryId?: number;
featured?: boolean;
available?: boolean;
};

export type ListDealsParams = {
active?: boolean;
};

export type ListOrdersParams = {
status?: ListOrdersStatus;
limit?: number;
};

export type ListOrdersStatus = typeof ListOrdersStatus[keyof typeof ListOrdersStatus];


export const ListOrdersStatus = {
  new: 'new',
  preparing: 'preparing',
  ready: 'ready',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;