# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Flare by TK Restaurant API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all menu categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a menu category (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "401":
          description: Unauthorized
  /categories/{id}:
    get:
      operationId: getCategory
      tags: [categories]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "404":
          description: Not found
    patch:
      operationId: updateCategory
      tags: [categories]
      summary: Update a category (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "401":
          description: Unauthorized
        "404":
          description: Not found
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete a category (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized

  # ── Menu Items ───────────────────────────────────────────────
  /menu-items:
    get:
      operationId: listMenuItems
      tags: [menu-items]
      summary: List all menu items, optionally filtered by category
      parameters:
        - name: categoryId
          in: query
          required: false
          schema:
            type: integer
        - name: featured
          in: query
          required: false
          schema:
            type: boolean
        - name: available
          in: query
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: List of menu items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MenuItem"
    post:
      operationId: createMenuItem
      tags: [menu-items]
      summary: Create a menu item (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MenuItemInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
        "401":
          description: Unauthorized
  /menu-items/{id}:
    get:
      operationId: getMenuItem
      tags: [menu-items]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Menu item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
        "404":
          description: Not found
    patch:
      operationId: updateMenuItem
      tags: [menu-items]
      summary: Update a menu item (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MenuItemUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
        "401":
          description: Unauthorized
        "404":
          description: Not found
    delete:
      operationId: deleteMenuItem
      tags: [menu-items]
      summary: Delete a menu item (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized

  # ── Deals ─────────────────────────────────────────────────────
  /deals:
    get:
      operationId: listDeals
      tags: [deals]
      summary: List all deals (active ones for customers)
      parameters:
        - name: active
          in: query
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: List of deals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Deal"
    post:
      operationId: createDeal
      tags: [deals]
      summary: Create a deal (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DealInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deal"
        "401":
          description: Unauthorized
  /deals/{id}:
    get:
      operationId: getDeal
      tags: [deals]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deal"
        "404":
          description: Not found
    patch:
      operationId: updateDeal
      tags: [deals]
      summary: Update a deal (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DealUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deal"
        "401":
          description: Unauthorized
        "404":
          description: Not found
    delete:
      operationId: deleteDeal
      tags: [deals]
      summary: Delete a deal (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized

  # ── Orders ────────────────────────────────────────────────────
...
components:
  schemas:
...
    Category:
      type: object
      required: [id, name, slug, sortOrder, isActive, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        sortOrder:
          type: integer
        isActive:
          type: boolean
        itemCount:
          type: integer
        createdAt:
          type: string
    CategoryInput:
      type: object
      required: [name, slug]
      properties:
        name:
          type: string
          minLength: 1
        slug:
          type: string
          minLength: 1
        description:
          type: string
        imageUrl:
          type: string
        sortOrder:
          type: integer
        isActive:
          type: boolean
    CategoryUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        slug:
          type: string
        description:
          type: string
        imageUrl:
          type: string
        sortOrder:
          type: integer
        isActive:
          type: boolean

    # ── Menu Items ─────────────────────────────────────────
    MenuItem:
      type: object
      required: [id, name, slug, categoryId, price, isAvailable, isFeatured, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        categoryId:
          type: integer
        categoryName:
          type: ["string", "null"]
        price:
          type: number
        originalPrice:
          type: ["number", "null"]
        imageUrl:
          type: ["string", "null"]
        isAvailable:
          type: boolean
        isFeatured:
          type: boolean
        calories:
          type: ["integer", "null"]
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
    MenuItemInput:
      type: object
      required: [name, slug, categoryId, price]
      properties:
        name:
          type: string
          minLength: 1
        slug:
          type: string
        description:
          type: string
        categoryId:
          type: integer
        price:
          type: number
        originalPrice:
          type: number
        imageUrl:
          type: string
        isAvailable:
          type: boolean
        isFeatured:
          type: boolean
        calories:
          type: integer
        tags:
          type: array
          items:
            type: string
    MenuItemUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        slug:
          type: string
        description:
          type: string
        categoryId:
          type: integer
        price:
          type: number
        originalPrice:
          type: number
        imageUrl:
          type: string
        isAvailable:
          type: boolean
        isFeatured:
          type: boolean
        calories:
          type: integer
        tags:
          type: array
          items:
            type: string

    # ── Deals ──────────────────────────────────────────────
    Deal:
      type: object
      required: [id, title, slug, discountType, isActive, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        discountType:
          type: string
          enum: [percentage, fixed, combo]
        discountValue:
          type: ["number", "null"]
        originalPrice:
          type: ["number", "null"]
        dealPrice:
          type: ["number", "null"]
        isActive:
          type: boolean
        sortOrder:
          type: integer
        createdAt:
          type: string
    DealInput:
      type: object
      required: [title, slug, discountType]
      properties:
        title:
          type: string
          minLength: 1
        slug:
          type: string
        description:
          type: string
        imageUrl:
          type: string
        discountType:
          type: string
          enum: [percentage, fixed, combo]
        discountValue:
          type: number
        originalPrice:
          type: number
        dealPrice:
          type: number
        isActive:
          type: boolean
        sortOrder:
          type: integer
    DealUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        slug:
          type: string
        description:
          type: string
        imageUrl:
          type: string
        discountType:
          type: string
          enum: [percentage, fixed, combo]
        discountValue:
          type: number
        originalPrice:
          type: number
        dealPrice:
          type: number
        isActive:
          type: boolean
        sortOrder:
          type: integer

    # ── Orders ─────────────────────────────────────────────
...