# lib/api-zod/src/generated/types/deal.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Flare by TK Restaurant API
 * OpenAPI spec version: 0.1.0
 */
import type { DealDiscountType } from './dealDiscountType';

export interface Deal {
  id: number;
  title: string;
  slug: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  discountType: DealDiscountType;
  /** @nullable */
  discountValue?: number | null;
  /** @nullable */
  originalPrice?: number | null;
  /** @nullable */
  dealPrice?: number | null;
  isActive: boolean;
  sortOrder?: number;
  createdAt: string;
}